### 2024-10-18 #7373 @kyleniemeyer
Hello @navasmontilla, we will use this issue to assign an editor and find reviewers.

Before assigning an editor, I noticed that your project's README is extremely long, and seems to contain essentially all the documentation. I strongly recommend that you reduce the length of the README to basic overview and use/installation instructions, and create a dedicated documentation website/webpages. This will almost certainly be a request from reviewers.
