### 2024-10-25 600e221
Create docExampleCases.md
