+++ doc/docExampleCases.md
@@ -10,71 +10,71 @@
 - Setting up a 1D linear transport case: [caseLinear](../python/caseLinear.py) ```python3 python/caseLinear.py```
   
   <figure style="text-align: center;">
-  <img src="doc/animationLin.gif" width="30%" alt="my alt text"/>
+  <img src="../doc/animationLin.gif" width="30%" alt="my alt text"/>
   </figure>
 
-- Setting up a 1D Burgers problem: [caseBurgers](python/caseBurgers.py) ```python3 python/caseBurgers.py```
+- Setting up a 1D Burgers problem: [caseBurgers](../python/caseBurgers.py) ```python3 python/caseBurgers.py```
   
   <figure style="text-align: center;">
-  <img src="doc/animationBurg.gif" width="30%" alt="my alt text"/>
+  <img src="../doc/animationBurg.gif" width="30%" alt="my alt text"/>
   </figure>
 
-- Setting up a 3D linear transport case: [caseLinear3D](python/caseLinear3D.py) ```python3 python/caseLinear3D.py```
+- Setting up a 3D linear transport case: [caseLinear3D](../python/caseLinear3D.py) ```python3 python/caseLinear3D.py```
   
   <figure style="text-align: center;">
-  <img src="doc/animationLin3D.gif" width="50%" alt="my alt text"/>
+  <img src="../doc/animationLin3D.gif" width="50%" alt="my alt text"/>
   </figure>
 
 ### Compressible flow (homogeneus Euler equations)
 
-- Setting up a 1D Riemann Problem for the  Euler equations (sod shock problem) [1]: [caseRP](python/caseRP.py) ```python3 python/caseRP.py```
+- Setting up a 1D Riemann Problem for the  Euler equations (sod shock problem) [1]: [caseRP](../python/caseRP.py) ```python3 python/caseRP.py```
 
   <figure style="text-align: center;">
-  <img src="doc/animationRP.gif" width="50%" alt="my alt text"/>
+  <img src="../doc/animationRP.gif" width="50%" alt="my alt text"/>
   </figure>
   
-- Setting up a 2D Riemann Problem for the Euler equations [2]: [caseRP2D](python/caseRP2D.py) ```python3 python/caseRP2D.py```
+- Setting up a 2D Riemann Problem for the Euler equations [2]: [caseRP2D](../python/caseRP2D.py) ```python3 python/caseRP2D.py```
 
   <figure style="text-align: center;">
-  <img src="doc/animationRP2D.gif" width="50%" alt="my alt text"/>
+  <img src="../doc/animationRP2D.gif" width="50%" alt="my alt text"/>
   </figure>
   
-- Setting up a 3D Riemann Problem for the Euler equations [3]: [caseRP3D](python/caseRP3D.py) ```python3 python/caseRP3D.py```
+- Setting up a 3D Riemann Problem for the Euler equations [3]: [caseRP3D](../python/caseRP3D.py) ```python3 python/caseRP3D.py```
 
   <figure style="text-align: center;">
-  <img src="doc/animationRP3D.gif" width="50%" alt="my alt text"/>
+  <img src="../doc/animationRP3D.gif" width="50%" alt="my alt text"/>
   </figure>
   
-- Advanced case: the shock-bubble test case [4]: [caseShockBub](python/caseShockBub.py) ```python3 python/caseShockBub.py```
+- Advanced case: the shock-bubble test case [4]: [caseShockBub](../python/caseShockBub.py) ```python3 python/caseShockBub.py```
 
   <figure style="text-align: center;">
-  <img src="doc/animationSB.gif" width="50%" alt="my alt text"/>
+  <img src="../doc/animationSB.gif" width="50%" alt="my alt text"/>
   </figure>
   
[…]
