+++ README.md
@@ -7,9 +7,6 @@
 2. [Installation](#installation)
 3. [Automated test](#automated-test)
 4. [Example usage](#example-usage)
-   - [Scalar problems](#scalar-problems)
-   - [Compressible flow (homogeneous Euler equations)](#compressible-flow-homogeneous-euler-equations)
-   - [Atmospheric flow (Euler equations with gravity source term)](#atmospheric-flow-euler-equations-with-gravity-source-term)
 5. [Functionality documentation](#functionality-documentation)
    - [Code organization and libraries](#code-organization-and-libraries)
    - [Configuration of the code for compilation](#configuration-of-the-code-for-compilation)
@@ -114,93 +111,8 @@ To get started in a user-friendly environment, some Jupyter Notebooks have been 
 
 For a correct functionality, *Jupyter Notebook* must be launched from the software main directory, e.g. ```me@myPc: SomeFolder/CAELUM/$ jupyter notebook```
 
-A more complete set of examples, scripted in Python, are listed below. The Python scripts to generate the cases below must be launched from the software main directory (```me@myPc: SomeFolder/CAELUM/$```).
+A more complete set of examples, scripted in Python, are can be found  [**here**](doc/docExampleCases.md). The Python scripts to generate the cases below must be launched from the software main directory (```me@myPc: SomeFolder/CAELUM/$```).
 
-### Scalar problems
-
-- Setting up a 1D linear transport case: [caseLinear](python/caseLinear.py) ```python3 python/caseLinear.py```
-  
-  <figure style="text-align: center;">
-  <img src="doc/animationLin.gif" width="30%" alt="my alt text"/>
-  </figure>
-
-- Setting up a 1D Burgers problem: [caseBurgers](python/caseBurgers.py) ```python3 python/caseBurgers.py```
-  
-  <figure style="text-align: center;">
-  <img src="doc/animationBurg.gif" width="30%" alt="my alt text"/>
-  </figure>
-
-- Setting up a 3D linear transport case: [caseLinear3D](python/caseLinear3D.py) ```python3 python/caseLinear3D.py```
-  
-  <figure style="text-align: center;">
-  <img src="doc/animationLin3D.gif" width="50%" alt="my alt text"/>
-  </figure>
-
-### Compressible flow (homogeneus Euler equations)
-
-- Setting up a 1D Riemann Problem for the  Euler equations (sod shock problem) [1]: [caseRP](python/caseRP.py) ```python3 python/caseRP.py```
-
-  <figure style="text-align: center;">
-  <img src="doc/animationRP.gif" width="50%" alt="my alt text"/>
-  </figure>
-  
-- Setting up a 2D Riemann Problem for the Euler equations [2]: [caseRP2D](python/caseRP2D.py) ```python3 python/caseRP2D.py```
-
-  <figure style="text-align: center;">
-  <img src="doc/animationRP2D.gif" width="50%" alt="my alt text"/>
-  </figure>
-  
-- Setting up a 3D Riemann Problem for the Euler equations [3]: [caseRP3D](python/caseRP3D.py) ```python3 python/caseRP3D.py```
-
-  <figure style="text-align: center;">
-  <img src="doc/animationRP3D.gif" width="50%" alt="my alt text"/>
-  </figure>
-  
-- Advanced case: the shock-bubble test case [4]: [caseShockBub](python/caseShockBub.py) ```python3 python/caseShockBub.py```
-
-  <figure style="text-align: center;">
-  <img src="doc/animationSB.gif" width="50%" alt="my alt text"/>
-  </figure>
-  
[…]
