### 2024-10-25 aa55217
Create docFunctionality.md
