+++ doc/docFunctionality.md
@@ -1,0 +1,458 @@
+## Functionality documentation
+
+- [Code organization and libraries](#code-organization-and-libraries)
+- [Configuration of the code for compilation](#configuration-of-the-code-for-compilation)
+- [Equations solved](#equations-solved)
+- [Input data](#input-data)
+- [Output data](#output-data)
+- [The computational mesh](#the-computational-mesh)
+- [Boundary conditions](#boundary-conditions)
+- [Spatial reconstructions](#spatial-reconstructions)
+- [Time integrator](#time-integrator)
+- [Riemann solvers](#riemann-solvers)
+
+### Code organization and libraries
+
+```
+caelum/
+├── main.c
+├── Makefile
+├── lib/
+│   ├── closures.c
+│   ├── closures.h
+│   ├── definitions.h
+│   ├── ibmutils.c
+│   ├── ibmutils.h
+│   ├── mathutils.c
+│   ├── mathutils.h
+│   ├── numcore.c
+│   ├── numcore.h
+│   ├── postproc.c
+│   ├── postproc.h
+│   ├── preproc.c
+│   ├── preproc.h
+│   ├── reconst.c
+│   ├── reconst.h
+│   ├── solvers.c
+│   ├── solvers.h
+│   └── structures.h
+├── python/
+│   ├── utils.py
+│   ├── autotest.py
+│   ├── caseExample.py
+│   └── caseExample.ipynb
+├── case/
+│   ├── configure.input
+│   ├── equilibrium.out
+│   ├── initial.out
+│   └── out/
+└── README.md
+```
+
+#### main Directory
+- **main.c**: Main source file for the CÆLUM solver.
+- **Makefile**: Makefile for compiling the code.
+- **lib/**: Directory containing library files and utilities.
+- **python/**: Directory containing Python scripts for pre/post-processing and automation.
+- **case/**: Directory for case configuration and output files.
+- **README.md**: Documentation for the CÆLUM solver.
+
+#### lib Directory
+- **definitions.h**: Definitions and constants used across the code.
+- **structures.h**: Data structures used in the code.
+- **closures.c/h**: Functions and headers related to pressure closures.
+- **ibmutils.c/h**: Utilities for immersed boundary method.
+- **mathutils.c/h**: Mathematical utilities.
+- **numcore.c/h**: Core numerical methods and routines.
+- **postproc.c/h**: Post-processing utilities.
+- **preproc.c/h**: Pre-processing utilities.
+- **reconst.c/h**: Reconstruction methods for high-order schemes.
+- **solvers.c/h**: Riemann solver routines.
+
+#### python Directory
+- **utils.py**: Utility functions.
+- **autotest.py**: Script for automated testing of the solver.
+- **caseExample.ipynb**: Script for generating case configurations.
+
+
+### Configuration of the code for compilation
+
+The file ```lib/definitions.h``` contains some definitions and constants that will be used for compilation. The most relevant for the user are listed below. Here, we provide a summary with links to other sections where some of these features are explained more in detail.
+
+#### Reconstruction Method
+
+```c
+#define TYPE_REC 0
+```
+
+- *Description*: Defines the type of high order reconstruction method used. See more info [here](#spatial-reconstructions).
+- *Possible Values*:
+  - `0`: WENO (Weighted Essentially Non-Oscillatory).
+  - `1`: TENO (Targeted Essentially Non-Oscillatory).
+  - `2`: Optimal Polynomial Reconstruction.
+
[…]
