+++ README.md
@@ -8,16 +8,6 @@
 3. [Automated test](#automated-test)
 4. [Example usage](#example-usage)
 5. [Functionality documentation](#functionality-documentation)
-   - [Code organization and libraries](#code-organization-and-libraries)
-   - [Configuration of the code for compilation](#configuration-of-the-code-for-compilation)
-   - [Equations solved](#equations-solved)
-   - [Input data](#input-data)
-   - [Output data](#output-data)
-   - [The computational mesh](#the-computational-mesh)
-   - [Boundary conditions](#boundary-conditions)
-   - [Spatial reconstructions](#spatial-reconstructions)
-   - [Time integrator](#time-integrator)
-   - [Riemann solvers](#riemann-solvers)
 6. [Other numerical results](#other-numerical-results)
 7. [Community guidelines](#community-guidelines)
 8. [Authorship](#authorship)
@@ -116,455 +106,9 @@ A more complete set of examples, scripted in Python, are can be found  [**here**
 
 ## Functionality documentation
 
+For additional information about the equations solved, code organization and libraries, input and output files, etc., see the [the functionality documentation](doc/docFunctionality.md)
+
 For a detailed documentation of the main programming structures and functions of the code, see [the API documentation](doc/docAPI.md)
-
-### Code organization and libraries
-
-```
-caelum/
-├── main.c
-├── Makefile
-├── lib/
-│   ├── closures.c
-│   ├── closures.h
-│   ├── definitions.h
-│   ├── ibmutils.c
-│   ├── ibmutils.h
-│   ├── mathutils.c
-│   ├── mathutils.h
-│   ├── numcore.c
-│   ├── numcore.h
-│   ├── postproc.c
-│   ├── postproc.h
-│   ├── preproc.c
-│   ├── preproc.h
-│   ├── reconst.c
-│   ├── reconst.h
-│   ├── solvers.c
-│   ├── solvers.h
-│   └── structures.h
-├── python/
-│   ├── utils.py
-│   ├── autotest.py
-│   ├── caseExample.py
-│   └── caseExample.ipynb
-├── case/
-│   ├── configure.input
-│   ├── equilibrium.out
-│   ├── initial.out
-│   └── out/
-└── README.md
-```
-
-#### main Directory
-- **main.c**: Main source file for the CÆLUM solver.
-- **Makefile**: Makefile for compiling the code.
-- **lib/**: Directory containing library files and utilities.
-- **python/**: Directory containing Python scripts for pre/post-processing and automation.
-- **case/**: Directory for case configuration and output files.
-- **README.md**: Documentation for the CÆLUM solver.
-
-#### lib Directory
-- **definitions.h**: Definitions and constants used across the code.
-- **structures.h**: Data structures used in the code.
-- **closures.c/h**: Functions and headers related to pressure closures.
-- **ibmutils.c/h**: Utilities for immersed boundary method.
-- **mathutils.c/h**: Mathematical utilities.
-- **numcore.c/h**: Core numerical methods and routines.
-- **postproc.c/h**: Post-processing utilities.
-- **preproc.c/h**: Pre-processing utilities.
-- **reconst.c/h**: Reconstruction methods for high-order schemes.
-- **solvers.c/h**: Riemann solver routines.
-
-#### python Directory
-- **utils.py**: Utility functions.
-- **autotest.py**: Script for automated testing of the solver.
-- **caseExample.ipynb**: Script for generating case configurations.
-
-
[…]
