### 2024-10-25 #7373 @navasmontilla
Dear @kyleniemeyer,

I am sorry for the delay in my reply. I could not upload the proposed changes earlier. Following your instructions, I have reduced the length of the README file. It now includes an introduction, installation/use instructions, and the community guidelines and license. I have moved the section containing the example cases and the functionality documentation to other files and included the links to them in the README file

Please let me know if it is ok in its present form.

Thank you very much for your assistance.

Adrián
