### 2025-02-13 #7373 @philipcardiff
Hello @Antonis-F, Would you be interested in reviewing the submission [CÆLUM: An academic High-Order Finite-Volume solver for the compressible Euler equations and related models](https://github.com/openjournals/joss-reviews/issues/7373), which we received at JOSS? The submission link is https://github.com/openjournals/joss-reviews/issues/7373, and the repository is at https://github.com/navasmontilla/CAELUM.

If you are unfamiliar with JOSS, the review process is unique: it takes place in a GitHub issue, is open, and encourages author-reviewer-editor conversations. To get an idea of what is expected, you can look at our review criteria (https://joss.readthedocs.io/en/latest/review_criteria.html) and review checklist (https://joss.readthedocs.io/en/latest/review_checklist.html).
