### 2025-02-14 #7777 @editorialbot
[REVIEW]: CÆLUM: An academic High-Order Finite-Volume solver for the compressible Euler equations and related models
**Submitting author:** @navasmontilla (<a href="http://orcid.org/0000-0002-3465-6898">Adrian Navas Montilla</a>)
**Repository:** https://github.com/navasmontilla/CAELUM
**Branch with paper.md** (empty if default branch):
**Version:** v1.0
