### 2025-02-14 #7373 @philipcardiff
I have started the review at https://github.com/openjournals/joss-reviews/issues/7777 and we can continue the discussion there.

If Antonis-F or other reviewers become available, I will add them also.
