### 2025-02-17 dc956c3
Update paper.bib
doi fixed
