+++ paper/paper.bib
@@ -300,7 +300,7 @@ keywords = {CFD, High-order, Finite-volume, Parallel, HPC, Open-source}
   number={3},
   pages={311--319},
   year={2015},
-  doi = {10.1177/10943420155768},
+  doi = {10.1177/1094342015576846},
   publisher={Sage Publications Sage UK: London, England}
 }
 
