### 2025-02-17 #7777 @navasmontilla
> Hi [@navasmontilla](https://github.com/navasmontilla), please fix the one invalid DOI noted above. Thanks.

Hi @philipcardiff , I have just fixed the invalid DOI. Thanks for your support.
