### 2025-06-02 navasmontilla/EHOW3D_public#1 @iBatistic
Automated Tests Failing
I encountered consistent failures when running the automated tests with various combinations of input parameters.

System Information:
	•	OS: Ubuntu 24.04.2
	•	Compiler: GCC 13.3.0

The log file is attached to this issue to help diagnose the problem.
Please let me know if additional details or test data would be helpful for reproducing the issue.

![Image](https://github.com/user-attachments/assets/6157eac5-4b99-48b7-a354-861c487507bb)

This is in the context of my review for the JOSS publication [https://github.com/openjournals/joss-reviews/issues/7777)](https://github.com/openjournals/joss-reviews/issues/7777#issuecomment-2658727832)

[log.dat.zip](https://github.com/user-attachments/files/20553271/log.dat.zip)
