### 2025-06-04 381a79f
Fixed sorting files for autotest
Fixed sorting files for autotest script. The output files are now read in ascending order.
