### 2025-06-04 navasmontilla/CAELUM#1 @navasmontilla
I tried, but I was not able to reproduce the failings in my PC. After examining the log file you shared above, I found out that the problem (I hope) is that the `glob` library in Python was not reading the output files in ascending order in your PC, whereas it was doing it in mine. Therefore, the metrics to decide wheter or not a test case is passed or fails were computed with the initial data (RP plots were representing the initial condition and the convergence plots where showing zero errors)

I have updated the `utils.py` file by adding an ordering instruction for the files before reading them: `files = sorted(glob(folder_out + "/*.out"), key=lambda x: int(re.findall(r'\d+', os.path.basename(x))[0]))` and I hope this fixes the problem.

I have also added the theoretical slopes in the convergence plot as you proposed.

These are the expected outputs for the following configurations:

-  `python3 python/autotest.py 1 0 3`

![Image](https://github.com/user-attachments/assets/dd39968c-2b70-4ebb-b1ea-038ab1b45c4f)

-  `python3 python/autotest.py 1 1 3`

![Image](https://github.com/user-attachments/assets/9e3b317f-c7ac-48ce-ae72-48f7b14884b4)

-  `python3 python/autotest.py 1 2 3`

![Image](https://github.com/user-attachments/assets/ee5da2e1-6cfe-42a2-b988-82bada690bdc)

When using the optimal reconstruction method (last case of the three above), Gibbs oscillations appear as a feature of this reconstruction method and therefore the RPs are not adequately solved (or the solution even explodes). This is why some RPs fail when setting the reconstruction to 2. This is what is expected from this type of reconstruction and motivates the use of WENO and TENO.
