### 2025-06-05 navasmontilla/CAELUM#1 @navasmontilla
To clarify the expected results, I have added this brief note in the *Automated Test* section of the `readme.md` file (main page of the repository):

**Note**: Test cases RP2, RP3, and RP4 are expected to fail when using the optimal reconstruction method (`rec=2`), as this method does not satisfy the esssentially non-oscillatory (ENO) property.

Please let me know if after yesterday's update of the Python library the tests run as expected.
