### 2025-06-05 7140d88
License info deleted from screen output
