+++ lib/preproc.c
@@ -357,6 +357,52 @@ int create_mesh(t_mesh *mesh, t_sim *sim){
 }
 
 
+int read_solid_cells(t_mesh *mesh, const char *folder_path){
+	int m,k,l,n,ct,tp;
+	FILE *fp;
+	char fname[1024];
+    
+    t_cell *cell;
+	cell=mesh->cell;
+	ct=0;
+      
+    snprintf(fname, sizeof(fname), "%s/solid_cells.input", folder_path);
+	fp = fopen(fname,"r");
+      
+	if (fp != NULL){
+		// Skip the first line
+		if (fscanf(fp, "%*[^\n]\n") != 0) {
+		  printf("Warning: Failed to skip the first line.\n");
+		}
+        if (fscanf(fp, "%*[^\n]\n") != 0) {
+		  printf("Warning: Failed to skip the first line.\n");
+		}
+
+		for(l=0;l<mesh->xcells;l++){
+			for(m=0;m<mesh->ycells;m++){
+				for(n=0;n<mesh->zcells;n++){
+				k = l + m*mesh->xcells + n*mesh->xcells*mesh->ycells;
+				if (fscanf(fp, "%*f %*f %*f %d", &tp) != 1) {
+				printf("%s Error: Failed to read solid cells data \n",WAR);
+				getchar();
+				}
+				cell[k].type=tp;
+				}
+			}
+		}
+
+		fclose(fp);
+
+		printf("%s solid_cells.input file has been read \n",OK);
+	}else{
+		printf("%s File solid_cells.input not found. \n",WAR);
+		ct=1;
+	}
+    
+    return ct;
+}
+
+
 int read_initial(t_mesh *mesh, t_sim *sim, const char *folder_path){
 	int m,k,l,n,ct;
 	FILE *fp;
@@ -408,6 +454,7 @@ int read_initial(t_mesh *mesh, t_sim *sim, const char *folder_path){
 				#else
 					gamma=_gamma_;
 				#endif
+                if(cell[k].type==1){
 				cell[k].Ue[0]=rho;
 				cell[k].Ue[1]=u*cell[k].U[0];
 				cell[k].Ue[2]=v*cell[k].U[0];
@@ -415,7 +462,16 @@ int read_initial(t_mesh *mesh, t_sim *sim, const char *folder_path){
 				cell[k].Ue[4]=energy_from_pressure(gamma,p,u,v,w,rho,cell[k].zc);
 				cell[k].Ue[5]=phi*rho;
 				cell[k].prese=p;
-				}
+				}else{
+                cell[k].U[0]=-1.0;
+				cell[k].U[1]=0.0;
+				cell[k].U[2]=0.0;
+				cell[k].U[3]=0.0;
+				cell[k].U[4]=0.0;
+				cell[k].U[5]=0.0;
+                cell[k].prese=0.0;
+                }
+                }
 			}
 		}
 
@@ -454,12 +510,21 @@ int read_initial(t_mesh *mesh, t_sim *sim, const char *folder_path){
 				#else
 					gamma=_gamma_;
 				#endif
+                if(cell[k].type==1){
 				cell[k].U[0]=rho;
 				cell[k].U[1]=u*cell[k].U[0];
 				cell[k].U[2]=v*cell[k].U[0];
 				cell[k].U[3]=w*cell[k].U[0];
 				cell[k].U[4]=energy_from_pressure(gamma,p,u,v,w,rho,cell[k].zc);
 				cell[k].U[5]=phi*rho;
+                }else{
+                cell[k].U[0]=-1.0;
+				cell[k].U[1]=0.0;
+				cell[k].U[2]=0.0;
+				cell[k].U[3]=0.0;
+				cell[k].U[4]=0.0;
+				cell[k].U[5]=0.0;
[…]
