### 2025-06-05 6477ec0
License info deleted from screen out
The previous commit was not adequate.
