### 2025-06-05 0119235
Cleaning license info from screen output
