+++ lib/preproc.c
@@ -1866,10 +1866,10 @@ void read_config(t_mesh *mesh, t_sim *sim, const char *folder_path){
 
 void print_info(t_mesh *mesh, t_sim *sim, const char *folder_path){
 
-      printf("\n\e[94mAuthors:\n - Adrián Navas Montilla\n - Isabel Echeverribar \n\n");
-      printf("Copyright (C) 2019-2024 The authors.   \n\n");
-      printf("License type: The 3-Clause BSD License \nRedistribution and use in source and binary forms, with or without modification, are permitted provided that the following conditions are met: \n 1. Redistributions of source code must retain the above copyright notice, this list of conditions and the following disclaimer. \n 2. Redistributions in binary form must reproduce the above copyright notice, this list of conditions and the following disclaimer in the documentation and/or other materials provided with the distribution. \n 3. Neither the name of the copyright holder nor the names of its contributors may be used to endorse or promote products derived from this software without specific prior written permission.\n\n");
-      printf("This software is provided by the copyright holders and contributors “as is” and any express or implied warranties, including, but not limited to, the implied warranties of merchantability and fitness for a particular purpose are disclaimed. In no event shall the copyright holder or contributors be liable for any direct, indirect, incidental, special, exemplary, or consequential damages (including, but not limited to, procurement of substitute goods or services; loss of use, data, or profits; or business interruption) however caused and on any theory of liability, whether in contract, strict liability, or tort (including negligence or otherwise) arising in any way out of the use of this software, even if advised of the possibility of such damage.\e[0m\n");
+      printf("\n\e[94m Authors:\n  - Adrián Navas Montilla\n  - Isabel Echeverribar \n");
+      printf(" Copyright (C) 2019-2024 The authors.   \n\n");
+      //printf("License type: The 3-Clause BSD License \nRedistribution and use in source and binary forms, with or without modification, are permitted provided that the following conditions are met: \n 1. Redistributions of source code must retain the above copyright notice, this list of conditions and the following disclaimer. \n 2. Redistributions in binary form must reproduce the above copyright notice, this list of conditions and the following disclaimer in the documentation and/or other materials provided with the distribution. \n 3. Neither the name of the copyright holder nor the names of its contributors may be used to endorse or promote products derived from this software without specific prior written permission.\n\n");
+      //printf("This software is provided by the copyright holders and contributors “as is” and any express or implied warranties, including, but not limited to, the implied warranties of merchantability and fitness for a particular purpose are disclaimed. In no event shall the copyright holder or contributors be liable for any direct, indirect, incidental, special, exemplary, or consequential damages (including, but not limited to, procurement of substitute goods or services; loss of use, data, or profits; or business interruption) however caused and on any theory of liability, whether in contract, strict liability, or tort (including negligence or otherwise) arising in any way out of the use of this software, even if advised of the possibility of such damage.\e[0m\n");
 
       printf(" \n");
       printf(" \e[4mSIMULATION SETUP:\e[0m\n");
