### 2025-06-06 #7777 @navasmontilla
Thanks @iBatistic @nasserma for your time testing the code. I have tried to address your comments about the autotest.py script in the repository [Issue #1](https://github.com/navasmontilla/CAELUM/issues/1).

Regarding the problems encountered by @iBatistic, I think they were caused by the `glob` Python library which was not sorting the files automatically in the post-processing of the results. I updated the `utils.py` script to solve it. I have also taken the oportunity to include the theoretical slopes in the convergence plot (image below), as well as to remove the license description from the screen output, as proposed.

![Image](https://github.com/user-attachments/assets/9bba01ed-808b-491d-9e10-3bd6051dafbe)

Regarding the observation raised by @nasserma, I would like to note that the optimal reconstruction method is a simple polynomial reconstruction without any limiters to handle discontinuities. Therefore, most RPs cannot be solved with this reconstruction method. This is why they fail when using the optimal reconstruction in the autotest. I have clarified this issue in the `readme.md` description of the automatic testing script. It would be an option not to run those RPs when setting the optimal reconstruction, but considering that the focus of this code is academic, I prefer to keep them so that we can show the limitations of such a simple reconstruction method and motivate the use of WENO and TENO.

I still have to address the other suggestions proposed by @iBatistic .
