### 2025-06-13 60da410
Update definitions.h
