### 2025-06-13 3ec7250
Updated libraries to include solid objects
Some libraries have been updated to include the definition of solid cells in the simulation.
