### 2025-06-13 053b85c
Update main.c
To include the definition of solid cells
