### 2025-06-13 d8f4d4c
updated python scripts to consider solids
Some python scripts have been updated to include the definition of solid cells.
