+++ python/caseDownwardBubbleSolid.py
@@ -1,0 +1,229 @@
+#!/usr/bin/env python
+# coding: utf-8
+
+# ## Configuration of a simulation case (*caseCollidingBub*)
+# 
+# ### Importing required libraries
+# 
+# Some of these are well known libraries such as *numpy* and *matplotlib* (they can be installed using pip). We also need to import the library *utils* containing predefined functionalities for this software. 
+
+import os
+import math                    
+import numpy as np             
+import matplotlib.pyplot as plt 
+import re
+from glob import glob
+from utils import modify_header_file,write_config,write_initial,write_equilibrium,backup_file,restore_file,compile_program,run_program,initialize_variables,read_data_euler,write_solid_cells
+import imageio
+
+# ### Setting up the paths
+# 
+# First, the name of the folder for this test case must be specified:
+
+#This test case will run in the folder "caseCollidingBub/". 
+#Don't forget the bar (/). 
+#This directory should have been created prior to the execution to this script, and should also contain an /out folder inside
+folder_case="caseDownwardBubbleSolid/" 
+
+
+# Then, all the paths are automatically assigned:
+
+
+#Do not modify the folders and paths below
+script_dir = os.path.dirname(os.path.abspath(__file__))
+folder_out="out/"
+folder_lib="lib"
+fname_config="configure.input"
+fname_eq="equilibrium.out"
+fname_ini="initial.out"
+fname_solids="solid_cells.input"
+folder_out = os.path.join(script_dir, "../"+folder_case+"/"+folder_out)
+folder_case = os.path.join(script_dir, "../"+folder_case)
+folder_lib = os.path.join(script_dir, "../"+folder_lib)
+folder_exe = os.path.join(script_dir, "../")
+
+os.makedirs(folder_case, exist_ok=True)
+os.makedirs(folder_out, exist_ok=True)
+
+for f in glob(folder_out + "/*.out") + glob(folder_out + "/*.vtk") + glob(folder_out + "/*.png"):
+    os.remove(f) 
+# ### Setting up the compilation variables (*definitions.h*)
+# 
+# Here, we can modify those variables that need to be set before compilation and are found in the file *definitions.h*. Don't worry if you mess up things here, a backup of the original file is created before modification and will be restored at the end of this script, after compilation and execution.
+
+#Do not change the line below, it creates a backup of the definitions.h file
+backup_file(folder_lib+'/definitions.h')
+#Configure the header file for compilation. Add as many lines as desired for the macros you want to modify.
+modify_header_file(folder_lib+'/definitions.h', 'NTHREADS', 48)         #number of threads
+modify_header_file(folder_lib+'/definitions.h', 'EQUATION_SYSTEM', 2)  #System of equations solved
+modify_header_file(folder_lib+'/definitions.h', 'ST', 3)               #Source term type
+modify_header_file(folder_lib+'/definitions.h', 'SOLVER', 0)           #Riemann solver used
+modify_header_file(folder_lib+'/definitions.h', 'READ_INITIAL', 1)     #Read or not initial data, this should ALWAYS be 1
[…]
