### 2025-06-13 #7777 @navasmontilla
Hi @iBatistic,

I have modified the convergence plot following your advice and it now looks as shown in the image below (WENO on top and optimal reconstruction below).

I have also made some small updates in the example cases (and related libraries) today.

Thanks for your time and help!

![Image](https://github.com/user-attachments/assets/f8b15b58-ac5c-4d54-8147-c4448c64f78d)
