### 2025-06-13 2f0e2be
Update docExampleCases.md
