### 2025-06-13 08d896d
Added examples with solid objects
