### 2025-06-13 af43829
Update docFunctionality.md
to include information about solid cells
