### 2025-07-07 4b92bf0
Update paper.bib
