+++ paper/paper.bib
@@ -271,7 +271,7 @@ keywords = {CFD, High-order, Finite-volume, Parallel, HPC, Open-source}
 }
 
 @article{kempf2024gal,
-  title={GAL $\{$$\backslash$AE$\}$ XI: Solving complex compressible flows with high-order discontinuous Galerkin methods on accelerator-based systems},
+  title={GALÆXI: Solving complex compressible flows with high-order discontinuous Galerkin methods on accelerator-based systems},
   author={Kempf, Daniel and Kurz, Marius and Blind, Marcel and Kopper, Patrick and Offenh{\"a}user, Philipp and Schwarz, Anna and Starr, Spencer and Keim, Jens and Beck, Andrea},
   journal={arXiv preprint arXiv:2404.12703},
   year={2024},
