### 2025-07-07 4d25082
added example cases description
according to review by I Batistic
