+++ paper/paper.md
@@ -71,7 +71,11 @@ Further details on the model equations and numerical resolution methods can be f
 
 # Previous work using the software
 
-`CÆLUM` was developed and first used in @NAVASMONTILLA2023 presenting novel methodology to construct very high order well-balanced schemes for the computation of the Euler equations with gravitational source term, with application to NWP. The objective of this work was twofold: first, to assess the use of augmented solvers in computing the Euler equations with gravity, and second, to evaluate the performance of the novel TENO reconstruction for NWP. Afterwards, a more traditional implementation of the Euler equations with gravity based on the use of fluctuation variables was implemented in different forms to also allow for the conservation of total energy. The resulting schemes were evaluated in terms of their spectral resolution for the computation of turbulent flows in the dry atmosphere. The results were presented in @NAVASMONTILLA2024, with the objective of determining whether or not these models can be used to build an iLES framework, shedding light on their potential advantages or limitations in representing under-resolved atmospheric processes in the meso- and micro-scales. 
+`CÆLUM` was developed and first used in @NAVASMONTILLA2023 presenting novel methodology to construct very high order well-balanced schemes for the computation of the Euler equations with gravitational source term, with application to NWP. The objective of this work was twofold: first, to assess the use of augmented solvers in computing the Euler equations with gravity, and second, to evaluate the performance of the novel TENO reconstruction for NWP. Afterwards, a more traditional implementation of the Euler equations with gravity based on the use of fluctuation variables was implemented in different forms to also allow for the conservation of total energy. The resulting schemes were evaluated in terms of their spectral resolution for the computation of turbulent flows in the dry atmosphere. The results were presented in @NAVASMONTILLA2024, with the objective of determining whether or not these models can be used to build an iLES framework, shedding light on their potential advantages or limitations in representing under-resolved atmospheric processes in the meso- and micro-scales.
+
+# Example cases
+
+The example cases provided showcase a diverse suite of test problems commonly used to validate and demonstrate the capabilities of computational fluid dynamics codes. These include scalar transport problems like linear advection and Burgers' equation in 1D, 2D and 3D, compressible flow simulations such as Riemann problems, shock–bubble interactions, and a Richtmyer–Meshkov instability for turbulent mixing. Additionally, atmospheric flow cases feature colliding thermal bubbles and descending thermal plumes over complex terrain. Each case is supported by Python scripts (or Jupyter notebooks) and corresponding visual animations, making them useful for verifying the numerical solvers and exploring the underlying physical phenomena. 
 
 
 # Acknowledgements
