### 2025-07-07 284382c
improved example cases section
according to review by I Batistic
