+++ paper/paper.md
@@ -75,7 +75,7 @@ Further details on the model equations and numerical resolution methods can be f
 
 # Example cases
 
-The example cases provided showcase a diverse suite of test problems commonly used to validate and demonstrate the capabilities of computational fluid dynamics codes. These include scalar transport problems like linear advection and Burgers' equation in 1D, 2D and 3D, compressible flow simulations such as Riemann problems, shock–bubble interactions, and a Richtmyer–Meshkov instability for turbulent mixing. Additionally, atmospheric flow cases feature colliding thermal bubbles and descending thermal plumes over complex terrain. Each case is supported by Python scripts (or Jupyter notebooks) and corresponding visual animations, making them useful for verifying the numerical solvers and exploring the underlying physical phenomena. 
+The example cases provided in the repository showcase a diverse suite of test problems that are commonly used to validate and demonstrate the capabilities of computational fluid dynamics codes. These include scalar problems like linear advection and Burgers' equation in 1D, 2D and 3D, compressible flow simulations such as Riemann problems (single- and multi-component flow), shock–bubble interactions, and a Richtmyer–Meshkov instability for turbulent mixing. Additionally, atmospheric flow cases feature colliding thermal bubbles and descending thermal plumes over complex terrain.  Each case is supported by Python scripts (or Jupyter notebooks) and corresponding visual animations, making them useful for verifying the numerical solvers and exploring the underlying physical phenomena. Apart from this, the code also provides an automated testing script that performs a convergence rate test, executes four Riemann problems, and runs an atmospheric flow benchmark.
 
 
 # Acknowledgements
