+++ .github/PULL_REQUEST_TEMPLATE/PR_template.md
@@ -1,0 +1,27 @@
+## 📝 Description
+
+Please provide a brief summary of the changes made in this pull request.
+
+## 🔧 Changes
+
+- [ ] Feature 1
+- [ ] Bugfix 2
+
+## 🧩 Related Issue(s)
+
+Related to #<issue_number>
+
+## ✅ Checklist
+
+- [ ] I have tested the changes locally
+- [ ] Documentation has been updated accordingly
+- [ ] The `autotest.py` script has been run and no errors are reported.
+- [ ] The pull request is ready for review
+
+## 📸 Screenshots (if needed)
+
+<!-- Add screenshots to help reviewers understand the changes -->
+
+## 📚 Additional Notes
+
+<!-- Any other information reviewers should be aware of -->
