### 2025-07-08 f6e6095
Update PR_template.md
improved template of the PR
