+++ .github/PULL_REQUEST_TEMPLATE/PR_template.md
@@ -2,10 +2,10 @@
 
 Please provide a brief summary of the changes made in this pull request.
 
-## 🔧 Changes
+## 🔧 Changes and improvements
 
-- [ ] Feature 1
-- [ ] Bugfix 2
+- [ ] The feature X has been added.
+- [ ] The bug X has been fixed.
 
 ## 🧩 Related Issue(s)
 
@@ -16,7 +16,6 @@ Related to #<issue_number>
 - [ ] I have tested the changes locally
 - [ ] Documentation has been updated accordingly
 - [ ] The `autotest.py` script has been run and no errors are reported.
-- [ ] The pull request is ready for review
 
 ## 📸 Screenshots (if needed)
 
