### 2025-07-08 aa7e01b
Update README.md
consulting and support info added
