### 2025-08-01 #7777 @iBatistic
Hi @navasmontilla; Okay, I’ll play with the other tutorial cases today and wrap up my review. Sorry for the wait!
