### 2025-08-05 #7777 @nasserma
Hello @navasmontilla, it seems as if most of the issues I found have been addressed, but there is one last significant issue regarding the paper. In the Statement of Need section, last paragraph, you seem to be trying to address the state of the field. I think this paragraph needs to be expanded to address two separate aspects of the state of the field in order to better understand the choices made in the development of CÆLUM:

1. the advantages/disadvantages of using higher-order FVM versus DG-FEM and hybrid DG-FEM
2. the a comparison of the specific attrributes of the package compared to existing open-source codes.

Regarding #2, you currently just list existing open-source codes and use a very high-level classification (FVM vs FEM), but I think more is required for this section.

Also, the paper needs to be proofed, there are a few spelling and formatting issues that were introduced in the last revision.
