### 2025-08-06 navasmontilla/CAELUM#2 @iBatistic
Missing xfvb - should be added in installation guide
This problem appears in the jupyter notebook cases

Solution: `sudo apt install libgl1 xvfb`

This is in the context of my review for the JOSS publication [https://github.com/openjournals/joss-reviews/issues/7777)](https://github.com/openjournals/joss-reviews/issues/7777#issuecomment-2658727832)
