### 2025-08-06 #7777 @iBatistic
Hi @navasmontilla,

I'm done with the review. I tried running the other cases you provided, and everything seems to be working (I opened a PR to add a missing library on Ubuntu for visualization).

As a reviewer, I noted that the authors followed the checklist required by JOSS and that everything necessary to fulfill the JOSS requirements was completed. I would like to emphasize to the authors that it would be beneficial to add more comments to the code, especially since it is presented as academic code. Although the amount of code is not large and is generally understandable, adding comments that directly link equations from the papers to the relevant parts of the code would make reading and understanding it much faster.

The code has already been used to publish two (or more) very good papers, and I hope the authors will continue to maintain and improve it. The code is written for a regular domain with a uniform structured mesh. Initially, I considered this a limitation, but for academic purposes and the intended applications, this approach is perfectly reasonable. As someone who works in a similar field, I understand the significant amount of time and effort required to develop such code from scratch.

p.s. I suggest changing the default case directories to CAELUM/run instead of CAELUM/.
