### 2025-08-07 #7777 @navasmontilla
Hi @iBatistic @nasserma,

thank you very much for your reviews. Throughout the revision process, your valuable and constructive feedback has significantly improved both the software and the paper, and I am truly grateful for your input. I would also like to thank you for pointing out some bugs or missing information needed for running the software adequately.

I will try to address these final comments as soon as possible (I will be 2 weeks out) and let you know when done.
