### 2025-09-29 #7777 @navasmontilla
Hi @philipcardiff ,

thanks for your help. I have done the following:

- Checked authors and affiliations.
- Generated new release (v1.1)
- Archived it in Zenodo with exactly the same name, authors (with Orcids) and license, with a link to the GitHub release. This is the archive: [https://doi.org/10.5281/zenodo.17224949](https://doi.org/10.5281/zenodo.17224949)

Let me know if everything is correct and if I need to follow additional steps. Thanks in advance!
